backend/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2022.12.7
distlib==0.3.7
Django==4.2.5
django-ckeditor==6.7.0
django-ckeditor-5==0.2.9
django-cors-headers==4.2.0
django-filter==23.3
django-js-asset==2.1.0
djangorestframework==3.14.0
filelock==3.12.4
Pillow==10.0.1
platformdirs==3.10.0
psycopg2-binary==2.9.7
pytz==2023.3.post1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.7.1
uWSGI==2.0.22
virtualenv==20.24.5
